refactor(trusty-common): split search_tui into submodules under 500-SLOC cap (#607)
The 1596-SLOC `search_tui.rs` is split into five focused submodules
under `search_tui/`:

- state.rs   (~170 SLOC)  — SearchTuiState + navigation methods
- nav.rs     (~90 SLOC)   — filter/sort, visible-ids, cursor movement
- render.rs  (~290 SLOC)  — row builders, stats/title lines, ratatui draw
- event_loop.rs (~220 SLOC) — daemon poll, reindex SSE, keyboard loop
- tests.rs   (~450 SLOC)  — all unit tests (test-file cap: 1500 SLOC)
- mod.rs     (~40 SLOC)   — thin re-export facade

All 219 tests pass; clippy -D warnings clean; fmt checked.
Removes the grandfathered entry from .line-cap-allowlist.tsv.
